% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listAirlines.R
\name{listAirlines}
\alias{listAirlines}
\title{Function that lists airlines retrieved from FlightStats}
\usage{
listAirlines(activeOnly = TRUE)
}
\arguments{
\item{activeOnly}{boolean to include all or only active airlines, can be set to TRUE or FALSE}
}
\value{
data.frame() with the airlines
}
\description{
Function that lists airlines retrieved from FlightStats
}
\examples{
listAirlines(activeOnly=FALSE)

}
\author{
Emelie Hofland, \email{emelie_hofland@hotmail.com}
}
