% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listAirports.R
\name{listAirports}
\alias{listAirports}
\title{Function that lists airports retrieved from FlightStats}
\usage{
listAirports(activeOnly = TRUE)
}
\arguments{
\item{activeOnly}{boolean to include all or only active airports, can be set to TRUE or FALSE}
}
\value{
data.frame() with the airlines
}
\description{
Function that lists airports retrieved from FlightStats
}
\examples{
listAirports()

}
\author{
Emelie Hofland, \email{emelie_hofland@hotmail.com}
}
