% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduledFlights.R
\name{scheduledFlights}
\alias{scheduledFlights}
\title{Function that retrieves flight schedules from FlightStats}
\usage{
scheduledFlights(airport, action, date, hour, silent = FALSE)
}
\arguments{
\item{airport}{character with the IATA code of the airport}

\item{action}{character, can be either arriving or departing}

\item{date}{character, a date in the format YYYY/MM/DD}

\item{hour}{character, hour in the format h or hh}

\item{silent}{boolean, to define if the number of flights found should be printed}
}
\value{
data.frame() with the flights for the hour specified
}
\description{
Function that retrieves flight schedules from FlightStats
}
\examples{
schedule <- scheduledFlights(airport="sxf",action = "arriving", date="2017/07/01", hour="10")

}
\author{
Emelie Hofland, \email{emelie_hofland@hotmail.com}
}
