% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduledFlightsFullDay.R
\name{scheduledFlightsFullDay}
\alias{scheduledFlightsFullDay}
\title{Function that retrieves flight schedules from FlightStats for the whole day}
\usage{
scheduledFlightsFullDay(airport, action, date)
}
\arguments{
\item{airport}{character with the IATA code of the airport}

\item{action}{character, can be either arriving or departing}

\item{date}{character, a date in the format YYYY/MM/DD}

\item{silent}{boolean, to define if the number of flights found should be printed}
}
\value{
data.frame() with the flights for the whole day
}
\description{
Function that retrieves flight schedules from FlightStats for the whole day
}
\examples{
schedule <- scheduledFlightsFullDay("sxf","departing","2017/07/01")

}
\author{
Emelie Hofland, \email{emelie_hofland@hotmail.com}
}
