% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduledFlightsFullWeek.R
\name{scheduledFlightsFullWeek}
\alias{scheduledFlightsFullWeek}
\title{Function that retrieves flight schedules from FlightStats for the whole week}
\usage{
scheduledFlightsFullWeek(airport, action, calendarweek, year = 2017)
}
\arguments{
\item{airport}{character with the IATA code of the airport}

\item{action}{character, can be either arriving or departing}

\item{calendarweek}{numeric, the calenderweek number of the week you want to retrieve flights for}

\item{year}{numeric, currently set default to 2017}
}
\value{
data.frame() with the flights for the whole week
}
\description{
Function that retrieves flight schedules from FlightStats for the whole week
}
\examples{
schedule <- scheduledFlightsFullWeek("sxf","departing",29)

}
\author{
Emelie Hofland, \email{emelie_hofland@hotmail.com}
}
