% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchAirline.R
\name{searchAirline}
\alias{searchAirline}
\title{Function searches a specific airline by IATA code}
\usage{
searchAirline(value)
}
\arguments{
\item{value}{character, an airline IATA code}
}
\value{
data.frame() with the airline
}
\description{
Function searches a specific airline by IATA code
}
\examples{
searchAirline("FR")

}
\author{
Emelie Hofland, \email{emelie_hofland@hotmail.com}
}
