% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchAirport.R
\name{searchAirport}
\alias{searchAirport}
\title{Function searches a specific airport(s) by IATA code, ICAO code or country code}
\usage{
searchAirport(by, value)
}
\arguments{
\item{by}{character, should be set to iata, icao or countryCode}

\item{value}{character, an iata, icao or country code}
}
\value{
data.frame() with the airport(s)
}
\description{
Function searches a specific airport(s) by IATA code, ICAO code or country code
}
\examples{
airports <- searchAirport(by="countryCode","NL")

}
\author{
Emelie Hofland, \email{emelie_hofland@hotmail.com}
}
