% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setAPIKey.R
\name{setAPIKey}
\alias{setAPIKey}
\title{Function to set the API key from FlightStats (developer.flightstats.com)}
\usage{
setAPIKey()
}
\description{
Function to set the API key from FlightStats (developer.flightstats.com)
}
\examples{
setAPIKey()
}
\author{
Emelie Hofland, \email{emelie_hofland@hotmail.com}
}
