% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setAppId.R
\name{setAppId}
\alias{setAppId}
\title{Function to set the AppID from FlightStats (developer.flightstats.com)}
\usage{
setAppId()
}
\description{
Function to set the AppID from FlightStats (developer.flightstats.com)
}
\examples{
setAppIdKey()
}
\author{
Emelie Hofland, \email{emelie_hofland@hotmail.com}
}
